+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Use [colr](https://github.com/welbornprod/colr) for colorization
 * Fix bug where checking for URL mismatch of untracked repos would cause an exception
 * Show more help when tracking no repos and no paths specified on command line
+* Note an error but continue gracefully when a repo doesn't have an origin
 
 ## [0.2.1] 2020-03-10